+++ catalyst_voices/lib/routes/guards/milestone_guard.dart
@@ -13,7 +13,7 @@ final class MilestoneGuard implements RouteGuard {
   @override
   FutureOr<String?> redirect(BuildContext context, GoRouterState state) {
     // allow milestone sub pages
-    if (state.uri.toString().startsWith('/$currentMilestone')) {
+    if (state.uri.toString().startsWith('/${Routes.currentMilestone}')) {
       return null;
     }
