+++ catalyst_voices/lib/routes/routes.dart
@@ -1,2 +1,3 @@
 export 'app_router.dart';
 export 'app_scaffold.dart';
+export 'routing/routing.dart';
